+++ projects/ion/src/lib/core/types/input.ts
@@ -1,0 +1,24 @@
+import { EventEmitter } from '@angular/core';
+import { SafeAny } from '../../utils/safe-any';
+
+export type IconDirection = 'left' | 'right';
+
+export type InputType = 'text' | 'password';
+
+export interface IonInputProps {
+  placeholder?: string;
+  button?: string;
+  iconInput?: string;
+  disabled?: boolean;
+  iconDirection?: IconDirection;
+  valid?: boolean;
+  invalid?: boolean;
+  inputButton?: boolean;
+  inputIconButton?: boolean;
+  clickButton?: EventEmitter<SafeAny>;
+  value?: string;
+  clearButton?: boolean;
+  inputType?: InputType;
+  valueChange?: EventEmitter<string>;
+  maxLength?: string | number | null;
+}
