+++ projects/ion/src/lib/input/input.component.ts
@@ -1,27 +1,6 @@
-import { SafeAny } from './../utils/safe-any';
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { IconType } from '../core/types/icon';
-
-export type IconDirection = 'left' | 'right';
-export type InputType = 'text' | 'password';
-
-export interface IonInputProps {
-  placeholder?: string;
-  button?: string;
-  iconInput?: string;
-  disabled?: boolean;
-  iconDirection?: IconDirection;
-  valid?: boolean;
-  invalid?: boolean;
-  inputButton?: boolean;
-  inputIconButton?: boolean;
-  clickButton?: EventEmitter<SafeAny>;
-  value?: string;
-  clearButton?: boolean;
-  inputType?: InputType;
-  valueChange?: EventEmitter<string>;
-  maxLength?: string | number | null;
-}
+import { IconDirection, InputType } from '../core/types/input';
 
 @Component({
   selector: 'ion-input',
